Generate rules with Python
This patch adds a Python script to generate the udev rules.  The
required data is defined in a TOML file.  This makes the rules easier to
maintain and update.

This causes some small changes in the formatting of the rules.  Also,
the rule for the Nitrokey Storage bootloader is updated to use ATTRS
instead of ATTR for consistency with similar rules.